Move chat controller into a module. Move websocket code into its own file
